COLIBRI control plane uses colibri paths.
Added more colibri debug commands: activate and cleanup.
Traceroute now checks that the requested reservation
starts at the local AS.
Client operator sets the destination IP to SvcCOL when C=1.
Store always passes a colibri path if available to use the operator.
Store has a patch function that returns the colibri path to a healthy
status after it lost the destination IA field. For that it uses the
steps inside the reservation.